ttools: fix FixSkyviewProjection.createAspect bug
The rotation matrix was not being used correctly when setting up
an aspect from a given sky position; the effect was that in some,
but not all, cases (CAR0 projection, or AITOFF/CAR projection
without reflected longitude) the SkyAspect got centered on the
wrong position.  That affected clat/clon setting in stilts plot2sky
and FOV settings in topcat.