ttools: add Aitoff0 sky projection alongside Aitoff
This is like the Aitoff projection except that the longitude=0
line appears at the left/right edge instead of down the center.

I implemented it a while ago but didn't offer it in the UI on
the grounds I wasn't sure it was useful.  But a couple of people
on the mailing list have requested this option recently,
so it looks like it is.