ttools: collapse whitespace in column descriptions
In the metadata display filter and output mode, collapse whitespace
in the column description before output; this prevents multi-line
descriptions (which can appear e.g. in VOTable DESCRIPTION elements)
from being displayed on multiple lines and generally makes their
display more sensible given that the output is assumed to be on
a single line of output.