ttools: add reverse cumulative option for histograms
The histogram-like plots that used to take a boolean isCumulative flag
now take a Cumulation object which can be NONE, FORWARD or BACKWARD.
A couple of people have requsted cumulative plots going in the
decreasing coordinate direction.

As it stands, this is a backwardly incompatible change, since plots
that used to supply the option cumulative=true/false must now write
cumulative=none/forward/reverse.  Perhaps I should try to mitigate that.