ttools: fix edge issue with cumulative histograms
When drawing a cumulative histogram, bars outside the range in which
samples were presented were not being plotted.  That's fine for a
non-cumulative histogram, but for cumulative ones it means that you
get an edge from maximum to zero at the right (or left) hand side
of the plot, which is not correct - the cumulative total stays at
the maximum all the way to infinity.  Looks especially wrong with
step-style bars.  Fix it.

Lorenzo Rimoldini pointed out this issue.