vo: fix column/fkey metadata display bug
Non-standard ("extra") column and foreign key metadata items are
supposed to be displayed in the relevant tabs of the TAP metadata
display.  Due to a bug, this wasn't working properly: the columns
appeared in the metadata tables, but their values weren't filled in,
at least for numerical data.  Fix this so they are displayed.