topcat: honour required datatype in JEL expressions
The TopcatJELEvaluator was ignoring the supplied required data type
and determining it automatically when compiling a JEL expression to
generate a RowAccess.
This was mostly fine, but failed when trying to read back in
certain saved sessions: synthetic columns with natural type Byte
get saved to FITS as type Short (because of the signed/unsigned
mismatch), so look like short when read back in from a saved session,
but when evaluated they yield Byte values from a Short column,
which causes trouble.

The user-visible effect (regression bug since the previous 4.8 release,
consequence of some of the multithreading enhancements) was big nasty
stack traces and blank values when trying to evaluate the column
in question (could be fixed by recompiling it with a few clicks in
the Columns window).

So the JEL expression has to evaluate to the requested type rather
than determining it on the basis of what the expression looks like.

I'm not sure why I didn't do it like that in the first place;
but I'm hoping it's because that caused something else to fail in an
obscure way which I couldn't track down at the time, caught by the
TopcatCodec unit test.  I've now fixed that (require JEL type
boolean.class not Boolean.class when compiling a SyntheticRowSubset
expression).  Hopefully there are no other nasties lurking that
this change will provoke.