topcat: fix JEL issue with type-changing columns
If synthetic column/subset A has an expression dependent on synthetic
column B, and B's expression gets reset in such a way that its
value changes data type, evaluation of A stops working.
The fix is to recompile A.  Add code in the column window which
now does this automatically.

Note it doesn't currently do it recursively, so if you really try
hard with nests of column expressions depending on each other and
then rewrite some column expressions to change their type,
you might still trigger a bug along these lines.
Cope with that if somebody reports it.