fits: avoid DataOutputStream.size method in fits-var output
The VariableFitsTableSerializer needs to keep track of the number
of bytes written to the heap in order to record data offsets.
It was doing this using the DataOutputStream.size() method.
That's fine for small files, but it returns an int, which maxes
out at Integer.MAX_VALUE when more than 2^31 bytes have been
written (arrgh - why would they do that??).  So calculate this
value by hand instead.

The effect was that variable-length array column values in fits-var
files beyond about 2Gb into the heap had bad values.