fits: P/Q columns can be read from streamed FITS
Modify the FitsTableBuilder so that input BINTABLEs with variable-length
array-valued columns (P/Q column descriptors) can be read even when
the input stream is not random-access (typically compressed data,
though maybe data from some non-file-based source as well).
This is generally problematic, because the cell data for such columns
is held in the heap after all the other cell data, so streaming
doesn't work.

To address this, the byte stream access is forced to provide random
access for such data sources, which is done by caching the entire
byte stream (under StorageProfile control) where required.