table: fix FileByteStore.copy large file bug
I misunderstood the usage of FileChannel.transferTo; multiple
invocations of this method may be required to transfer a given
amount of data.  I was only invoking it once, which meant that
in principle a FileByteStore might not restore all the data it
has accumulated.  This long-standing issue could have broken a
lot of things, but in practice it looks like it only happened
for (some?) files >2Gb.

Fixed now.