ttools: fix and enhance tcube task
TableCube wasn't handling input filters properly in the case that
they were non-random-access and there was no scale value set.

Fix that up, and take the opportunity at the same time to fix it
so that the cols and scale parameters take JEL expressions rather
than just column name/ID values.  There is a small risk of
backwards incompatibility here, but it's not very likely that it
will break actual working code.