parquet: fail gracefully in absence of parquet-mr
Fix the STIL parquet I/O handlers so that if the parquet-mr libraries
are not present and correct, they will report an error to that effect
rather than throw a caught or uncaught error at initialisation time.
The input handler can still detect parquet files by magic number.

This may be required since the bloated and special-interest parquet-mr
jar files may not always be packaged with these classes.