parquet: provide single dependencies jar file
Replace the multiple dependencies of the PARQUET library with
a single item, parquet-mr-stil.jar.  This file has no content except
for further Class-Path entries pointing to the actual parquet-mr
dependency libraries.  Doing it like this makes it easier to manage
packaging options that want to omit the (large and special-interest)
parquet-mr dependency jars.