fits: avoid THEAP in fits-var output
The VariableFitsTableSerializer was block-aligning the heap by
adding padding bytes between the end of the table data and the
heap data, with the size of this padding encoded in the THEAP
keyword.  This is legal FITS (and used in the example header)
as defined by sec 7.3.5 of FITS v4.0, but causes problems with
some other FITS reading software.

The fverify tool in FTOOLS erroneously reports such files as invalid,
since it misinterprets the value of the PCOUNT keyword as the
size of the heap alone, rather than the heap size plus pre-heap gap.
This was confirmed, after some resistance, by Bill Pence for the
NASA FITS support office, ticket FTOOLS #4219.
This bug is apparently present for at least FTOOLS versions
v3.14-v3.50.

In what looks like a separate but similar bug, astropy.io.fits
fails to read the same files.

Following advice from Bill Pence:

   "BTW, the THEAP keyword is not required and is rarely used by other
    projects. Unless you have a good reason to do so, I would recommend
    not using it. In that case, the heap will start immediately
    following the fixed-length table."

this commit modifies STIL so that it no longer writes pre-heap
padding, and no longer writes the THEAP header - said heading wasn't
really necessary, it just seemed like the tidy and FITSy thing to do.
The resulting output files seem to work OK with fverify and
astropy.io.fits.

This commit also includes for reference a test file theap-gap.fits
which triggers the offending behaviour, it was produced by running
in STILTS v3.4:

   stilts tpipe in=:loop:500
                cmd='addcol arr slice(intArray(0,1,2,3,4),0,i%6)'
                ofmt='fits-var(date=false)' out=theap-gap.fits

When run through fitsverify from CFITSIO 3.340 it causes the message:
"*** Error:   checking data fill: tried to move past end of file";
and the following astropy:
>>> from astropy.io import fits
>>> f = fits.open("theap-gap.fits")
>>> f[1].data
reports "OSError: Could not find heap data for the 'arr'
variable-length array column.".