tjoin: cds-healpix-java bugfix update to v0.29.2
Upgrade the cds-healpix-java library to v0.29.2.

This avoids a numeric bug that could generate an
ArrayIndexOutOfBoundsException for certain rare parameter values
when calculating cells overlapping a cone.
This could cause sky crossmatches to terminate with an error.

This is in fact a resurfacing of about the same bug that provoked
the last cds-healpix-java upgrade, but the fix this time is more robust
(the data structure in question has now been made elastic rather
than just increasing its maximum size as before), so it should not
crop up again.
See https://github.com/cds-astro/cds-healpix-java/issues/15.

A unit test is added to check that the current library version
has fixed this bug.

The jar file was built from source, using the sources from
https://github.com/cds-astro/cds-healpix-java,
at tag v0.29.2 (743a57bdd2d6a).