table: fix ASCII format type inference logic bug
Fix a logic error in the way that ASCII-like formats such as CSV
assess cell string values to determine the data type of the column.
Up till now, it was assuming that if a column looked like a boolean,
it shouldn't be assessed for looking like an integer and so on.
Instead, each datatype compatibility should be assessed independently
of all the others.

This long-standing bit of logic in the RowEvaluator looked pretty
confused, but it has apparently worked for the CSV files that
people have thrown at it for more than a decade, finally defeated
by a column that started "F,F,F" and carried on "1,1,1".
I hope that the replacement logic, which I'm pretty sure makes more
sense than what was there before, doesn't break anything.