table: LineSequence works with CR or CRLF
Modify LineSequence so that its output is identical for input
files with DOS-style line endings (\r\n) or Unix-style
line endings (\n).  This class is currently only used for IPAC
format files, and it doesn't seem to make a difference in that
case (the IPAC reader probably strips whitespace anyway),
but this makes it a bit more robust.