fits: read empty string in 1A column as null not '\0'
The Character result of reading a zero byte from a FITS column
representing a single character (TFORMn='1A') is now a java null
rather than a Character.valueOf(0) (ASCII NUL).
This makes sense, since a FITS string may be terminated before
the end by a \0, and it's healthier since you don't end up with
non-printable '\0' characters for output.

Most of the FITS output handlers already encode empty strings
or null Characters in this way (see later commits to ensure that
for all of the FITS output handlers).