fits: add allowZeroLengthString config option
Add an option to the FITS output handlers to prevent writing of
zero-length string columns, i.e. with TFORMn='0A'.
Such columns are explicitly permitted by the FITS standard,
but current CFITSIO versions (v3.50 and earlier) choke on them,
for instance causing a core dump in fitsverify, and apparently
trouble in (FTOOLS) fv and ftlist too - see FTOOLS ticket #4231.
This was causing problems with GACS output.

Default behaviour is unchanged, i.e. zero-length string columns
will be written where all strings have zero length.

This option is currently not documented at the user level
(there is no @ConfigMethod annotation on
AbstractFitsTableWriter.setAllowZeroLengthString).
It would be easy to do that, but it would clutter up the user
documentation with an option that most people probably don't need.
It is possible in any case to use this from the command line,
with e.g. a STILTS parameter like
"ofmt=fits-basic(allowZeroLengthString=false)".