ttools: relax taplint TAP 1.0 TAP_SCHEMA col type checks
Modify taplint checking of the columns in TAP 1.0 services.
Since TAP 1.0 has no real type system (see TAP 1.0 Erratum #3)
any apparent mismatches can't be real errors, so downgrade
the relevant reports from E_TSCT to W_TSCT and modify the wording.

Also, modify the way these tests are done.  Previously, the
(necessarily sloppy) matching was done between declared TAP 1.0
types and the required types "varchar" and "integer" as
written in TAP 1.0 sec 2.6.  However, the compability checking
such as it is is case-sensitive, based on the types listed in
the table in section 2.5, for which the relevant values are
"VARCHAR" and "INTEGER" instead.  So test against these upper-case
types instead.

The effect should be to eliminate or downgrade to Warning some
E-TMS-TSCT reports (e.g. IRSA).