fits: write TZERO cards as numeric, not strings
The TZEROn header card is supposed to be a numeric value, not
syntactically a string, and fitsverify complains if it's not.
Up till now it's been output wrapped in quotes, hence syntactically
as a string value, because of the difficulty of getting nom.tam.fits
to do otherwise.  But hack round it here so that it actually works.

The hack is somewhat messy, but hopefully it will be robust enough.
If we switch to using a later version of nom.tam.fits I can
use the relevant methods available there instead.