fits: string padding now configurable, defaults to '\0'
Modify the way that undersized (including zero-length) character array
data is written to TFORM=nA columns (for n>=1).

Up till now, those values were in some cases padded to length with
space characters.  Although that seemed like a FORTRAN/FITS-friendly
way to do things, I can't see anything in the FITS standard that requires
it, so the behaviour is now changed to pad, by default, with ASCII NUL
characters ('\0').  According to the FITS standard (FITS v4.0,
section 7.3.3.1) the first such NUL is encountered as a string
termination character, so that under-length strings can be read
as intended.

Simultaneously, this behaviour has become configurable, controlled
by the setPadCharacter method on AbstractFitsTableWriter
(via FitsTableSerializerConfig).  This is not currently documented
at the user level (with a @ConfigMethod annotation); that could
easily be done, but I don't expect people will want to be bothered
with this.  But it's there to revert to the previous behaviour
if that becomes necesssary.