table: fix drag'n'drop regression failure
Fix a bug that meant the STIL table drag'n'drop wasn't working;
the StarTableOutput was failing to set up a suitable table
serializer at construction time, as a consequence of some
changes that were made when bean-configurable I/O handlers
were introduced.  This was a regression bug since the last release,
noticed by a user who missed this not-very-well-known functionality
before I did.

Fix it so that a suitable handler (now VOTable/BINARY2 rather than
VOTable/BINARY) is set up, and put logging and unit tests in place
so that if this setup fails again in the future it will be noticed.