ttools: improve votlint message counts using codes
The PrintSaxMessager which is used to generate output for the
votlint application now looks at the VotLintCodes supplied with
messages rather than their text to determine whether the
same-message repeat count has been reached.  This is smarter,
and should result in more compact output in some cases
(e.g. messages with text content that is similar but not identical).

An info message is reported when repeated messages are suppressed.
It should have been doing something like that before, but didn't.