ttools: taplint downgrade some W_EXVC to I_EXVT
In case the DALI Examples RDFa @vocab attribute is correct but
known to cause trouble with some TOPCAT versions (v4.3-4.4),
downgrade the report from Warning to Info.  Semantically this
really is more of a Warning than an Info, but some people don't
like to see Warnings they can't get rid of, and the impact is now
quite low since the relevant topcat versions are quite old (2017).