table: fix broken jdbc URL regression bug
When I reworked JDBC handling in StarTableFactory using the TableScheme
framework, I broke it - the "jdbc:" prefix on JDBC URLs got lost
when presenting a ":jdbc:..." or "jdbc:..." table location, and the
load failed with an IllegalArgumentException ("Not a JDBC-protocol URL").
This affected supplying JDBC URLs by hand or by the SQL Query window
in topcat.  The error didn't get noticed because JDBC testing is not
included in the unit tests, because of the difficulty of having an
RDBMS on hand.  Given that this didn't get noticed for some months
following release, it looks like not many people are using direct
JDBC interaction (I'm not very surprised).  Fixed now.

Issue was reported by Jochen Liske.