topcat: build-standalone adds topcat-extra.jar (parquet)
Modify the build-standalone ant target so that it builds a new
superjar topcat-extra.jar, alongside the existing topcat-full.jar
and topcat-lite.jar.  At present, this is just like topcat-full,
but it contains the (large) parquet support libraries as well.

This organisation might be reviewed in the future; maybe the parquet
libs will be included in topcat-full if they get popular, or maybe
topcat-extra will be extended to include other special-interest bloat.