topcat: fix superjars to include top-level files
When I modified the way the superjars (topcat-full.jar etc) were
built to exclude a load of junk from the parquet-mr jars,
I inadvertently threw out the few flat files deliberately inserted
at the top level of the output jar file (scripts and licence file).

Reinstate them here.  It's a bit messy, since they have to be
inserted individually for each output superjar (there are 3 now),
but it works, and it's not easy to do with the existing options
in the SuperJar utility without bringing back the parquet files.