ttools: add tests for EpnTap standard
Add tests in taplint.DmTest unit test to ensure that the UCDs
and Units prescribed by the EPNTAP standard are legal;
this is a test of the EPNTAP standard more than of the taplint
code (or of any service).

Actually they are not all legal: at present complaints are simply
written to stderr at test time.  These will be raised with the
EPN-TAP authors.