vo: update operational UserAgent strings as per SoftID
Update the "(IVOA-xxx)" operational identifers applied to
operational client (e.g. validator) User-Agent HTTP headers,
according to the content of the published SoftID IVOA Note.
This differs from the draft content in earlier unpublished
versions and on the UserAgentUsage IVOA wiki page:
"(IVOA-validate)" is replaced by "(IVOA-test)".

Use of this utility in taplint and datalinklint are also
necessarily updated.