util: remove dependency on apache-bzip2.jar
The BZIP2 de/coding support classes are now compiled from source
in uk.ac.starlink.util.bzip2, so the additional dependency is
no longer required.

This is not a big deal for the normal build, but it means that
the POM dependency of STIL on ant-1.7.0 can be removed,
saving >1Mb of mostly irrelevant library classes.

At this commit the testcases have not switched to use the new classes.
Although the org.apache.tools.bzip2 classes are no longer an
explicit dependency, they are present on the classpath at test time
as part of the ant environment.  Running the tests with the old
classes against the application code using the new classes
constitutes a test that the two codesets are doing the same thing,
so the unit tests at this commit should (and do) run correctly.