ttools: avoid multithreaded high-order healpix plots
Revert to sequential processing in the HealpixPlotter when
HEALPix view level is high enough to cause resource usage issues
(per-thread copies of data structures whose size scales as level**4).
The details are ad hoc, but it improves performance for at least
some cases (e.g. level=11, 4Gb heap, 8 threads, which was grinding
to a halt with GC thrashing before).