ttools: add UCD and VOUnit validation harnesses
Add new classes UcdStatus and UnitStatus.  These are basically
adapters for Gregory Mantelet's Ucidy and Norman Gray's Unity parsers,
but some additional work is done to present the results in a way
that's easy to use in various places where they will be used in ttools,
and provide a bit more logic in places, for instance to accommodate
UCD1s as well as UCD1+s.