table: ASCII formats treat all-blank columns as String
Up till now, if the RowEvaluator (used for CSV parsing etc) encountered
a column with all blank values, it reported it as a Boolean column.
This isn't really helpful, especially if the table makes its way
into topcat where such columns end up as (useless) subsets.
So report it instead as a zero-length String column.