ttools: upgrade ucidy 1.1 -> 1.2-beta
This upgraded ucidy now validates against EN-UCD1+-1.4;
the previous version was with reference to REC-UCD1+-1.3.

There are some other useful changes: the jar file manifest no longer
contains a "." entry in the Class-Path which was causing problems
with my standalone jar file builds, and the default UCDParser instance
is no longer protected which makes it more straightforward to write
custom instances.

This ucidy jar file was downloaded from github; it corresponds to
2dc4cdd1897 at https://github.com/gmantele/ucidy.
The beta/pre-release status is because of an unresolved issue about
deprecation of meta.ref.ivorn in the standard.