ttools: update ucidy again
New version deals with a couple of issues: avoid output direct to stderr,
and put the wordlist resources in a place (not the top level of the
package namespace) where they don't interfere with superjar creation.

This ucidy jar file was downloaded from github; it corresponds to
a831389d01 at https://github.com/gmantele/ucidy.