table: fix warning count bug in Codec
Fix a less serious bug that could manifest itself when doing parallel
processing of large cached tables; the warning count is shared
between threads, so if the stream is corrupted the warnings might
get output the wrong number of times.  This code is not much
encountered though, and the output warning count limit is not
critical, so it wouldn't have caused too much trouble.