topcat: fix row sort error in 2 activation actions
There was some confusion in ActivationType/Activator implementations
about the meaning of the row index parameter to Activator.activateRow.
It's actually the index into the TopcatModel DataModel, not the
Apparent Table, so row sorting has been applied.
InvokeDataActivationType and ServiceActivationType were operating
on apparent table snapshots, so for sorted tables were picking up
the wrong row.

Fix this for the two activators in question, and add clarification
to the Activator javadocs.

This bug was reported by Yannick Roehlly.