topcat: add auto-invoke option to DatalinkPanel
The DatalinkPanel now has a hasAutoInvoke option that can be specified
at construction time.  That causes it to present an Auto-Invoke toggle
button which makes the selected datalink link automatically followed
whenever the panel is configured with a new table.

At this commit, the new functionality is not used anywhere.