ttools: handle foreign namespaces better in votlint
Fix votlint so that if it encounters an element in a non-VOTable
namespace it provides more nuanced reporting.
If it looks like a VOTable element in a non-VOTable namespace it
says one thing, but if it looks like a really foreign element
(e.g. some VODML) it says something else.

Eventually VODML should probably be handled explicitly as a known
namespace including content with known rules (though it's going to
be a bit tricky if VODML is defined in XSD 1.1, not generally
supported by Java 8).