tjoin: rework LinkSet usage in RowMatcher
Modify the table creation methods in RowMatcher so that they take
Collection<RowLink> objects, which are optionally ordered,
rather than LinkSet objects which are now documented as unordered.
Instead of the LinkSet.sort() method, which sorted the linkset in place,
there is now a method LinkSet.toSorted(), which yields a
Collection<RowLink>.

The purpose of all this is to keep separate operations on the links
that may modify them (these are now all operations on LinkSets,
and take no notice of ordering) from operations which just read
RowLink objects (these are now operations on, optionally ordered,
Collection<RowLink> objects).  That provides two benefits over
the previous arrangement.  Firstly modifying operations are always
done on Sets, so should be relatively cheap (previously in some
cases I was calling potentially many remove() operations on ArrayLists,
which could be quite expensive).  Secondly, the sorting is done
more transparently rather than as an implementation detail,
which means it can be timed, monitored, and done in parallel
or not done at all.