ttools: avoid use of plot.Mark{Shape,Style} class in plot2
The classes

   uk.ac.starlink.ttools.plot.MarkShape
   uk.ac.starlink.ttools.plot.MarkStyle

are relics of the old-style plotting; they contain baggage unnecessary
for the plot2 framework and use some classes that I'd like to get rid of
outside of the legacy uk.ac.starlink.ttools.plot package.
This commit replaces them throughout the plot2 framework with the
mostly equivalent classes

   uk.ac.starlink.ttools.plot2.layer.MarkerShape
   uk.ac.starlink.ttools.plot2.layer.MarkerStyle

MarkStyle creation may be slightly slower, but this is only done
a few times per plot by the TOPCAT GUI, not per point/row.
Clipped pixel iterator creation, which is done per point/row,
may be a bit faster.  But probably neither of these effects
will be noticeable.