ttools: avoid use of plot.Drawing class in plot2
Replace use of uk.ac.starlink.ttools.plot.Drawing with new class
uk.ac.starlink.ttools.plot2.PixelDrawing.
This has the same functionality and basically the same implementation,
so should be no change in behaviour for now, but some minor efficiency
improvements and the capability to extend the functionality in future
without messing with the legacy ttools.plot classes.