ttools: fix bug in partial range specifications
There was a regression bug in BasicRanger (related to NaN handling,
and introduced with Spans) that meant specifying ranges with a
lower finite limit worked, but with an upper finite limit didn't.
That meant for instance that a plot with auxmin but no auxmax worked OK
(plot used supplied lower limit but worked out upper limit from the data),
but auxmax with no auxmin did not (the supplied upper limit was ignored).
Reported by Sergei Klioner.  Now fixed.