ttools: avoid use of plot.ErrorRenderer in plot2
Replace the plot.ErrorRenderer class with plot2.layer.MultiPointShape;
much of the code is copied, but there are some changes in the
implementation.  In particular the plot.Drawing class is avoided,
and PixerFactories (mostly PixelDrawings) are used throughout.
The PixelDrawing class can be modified later without potential
consequences for the plot.* classes.

Additional code in config.MultiPointConfigKey was largely copied
from ttools.gui.MarkStyleSelectors.

Plot output ought not to be changed at this commit, though
performance characteristics might be (hopefully for the better).
The unit tests including pixel-for-pixel comparison of test plots
back this up.  However, the implementation changes are quite
extensive, so unexpected changes are not impossible.