ttools: improve cap/dart thick shapes
Modify the MultiPointShape so that when thicker lines are drawn,
those shapes which need it are adjusted accordingly before the
lines are thickened.  For instance an arrow needs a bigger head
if it's been drawn with a fatter pen otherwise the barbs get engulfed.
This affects (some) CappedLine and Dart shapes.