ttools: improve UCD1 recognition in UCD parser
The UCD parser now recognises lower-case UCD1s as well as upper case
ones.  It's not clear whether UCD1s are supposed to be case-sensitive
or not (there doesn't seem to be any documentation of them), but if
they've been through RegTAP ingestion they will be lower case anyway,
so better recognise them in that form.

Also treat the small number of tokens which are legal as both UCD1
and UCD1+ as UCD1+ instead of UCD1.  A special test is introduced
for this case.