ttools: DatalinkValidator checks SD name/description
Datalink 1.1 sec 4.1 recommends supplying a name attribute and
DESCRIPTION child for service descriptors "if the semantics are
not obvious,  and especially in the case of multiple sibling
service descriptors, or non-standard services".

So issue a Warning if there are multiple service descriptors
and they don't all have (unique) names and descriptions.

This commit is based on PR-DataLink-1.1-20230413.