ttools: refactor DatalinkValidator semantics checking
The terms in the semantics columns are now checked against the
datalink_core vocabulary using the VocabChecker class.
This provides more capable checking since it can check against
run-time updates of the vocabulary that were not available at
compile time.

This change also provides a framework that makes it easier to
check other vocabulary-constrained values.