ttools: adjust SVG graphics output
Following advice from Markus Demleitner, informed by an article at
https://css-tricks.com/scale-svg/,
add a viewBox attribute to the svg element written by SVG graphics
output.  This is, as far as I can tell, recommended behaviour,
and it seems to improve image scaling in some contexts.
In particular when the SVG is referenced from an HTML OBJECT element
(not just from an IMG) with style="width:100%" it does indeed scale up
to page width, while retaining its preferred size in pixels when
no width is specified.  This looks like a bit of a minefield though,
so it's possible that future requirements/experience will mean this
needs further adjustment.

At the same time, modified the output so that it contains just the
svg element itself, with no XML or DOCTYPE decoration.
I haven't had a request for this, but I expect it will make it
easier to re-use/embed this output in other XML content, while
not impacting its use when referenced from HTML elements.