topcat: fix inverse algebraic subset session load bug
The TopcatCodec2 session load operation was not correctly restoring
InverseRowSubsets based on SyntheticRowSubsets - they ended up
apparently containing all rows in the table.  The problem was
that the InverseRowSubset was based on a dummy subset before its
SyntheticRowSubset had been properly restored.

Now fixed; the problem was with load not save, so any sessions
saved can be loaded correctly with the fixed code.

Adjusted the unit tests so it would pick up this problem if still present.