topcat: fix another session restore bug
Fix a session save/restore bug that only showed up if an inverse
subset was defined, and at least two other subsets defined before
it had been deleted before saving.  The problem was that the
deleted subsets are serialized to the session file with the
placeholder name "DELETED" (that was a bad decision, but changing it
now wouldn't be sufficient since old session files are out there),
and that means that they don't get added separately when the session
is restored, since only one subset with a given name is permitted.
That meant that the subset restoration got confused about which subsets
appeared at which index, before it re-deleted the deleted ones again.
Grrr.  Fixed now by hacking the subset names to be unique on load.