topcat: auto-invoke datalink rows synchronously
Modify the way that the ViewDatalinkActivationType and (indirectly)
the InvokeDatalinkActivationType activation actions cause auto-invoked
actions to be invoked.  They are now done synchronously not
asynchronously.  That means that a more appropriate Outcome can
be logged in the activation window, and also that the auto-play
feature works properly; previously auto-playing these activation
actions had the effect of cancelling most of the actions before
they had been completed.

In implementation terms, what's happened is that the auto-invoke
is now just a UI-visible flag in the DatalinkPanel;
the actual activation is no longer done automatically as a
consequence of updating the DatalinkPanel state, it has to be done
explicitly by the calling code (the Activator).
This is a much more sensible way of doing it.