topcat: avoid unnecessary subset recounts
If the expression of a synthetic Row Subset is selected for editing
but not actually changed, do not trigger a recount of the rows it
contains.  Such selection can happen if the user just wants to
copy the expression to the clipboard for later use, and the recount
may be expensive.
This change suggested by Sergei Klioner.