ttools: fix MOC-ASCII parse error
If ASCII MOC strings presented for area plotting included a
depth specifier, which looks like a cell specifier with no cell part,
it was interpreted as a cell index (e.g. token "3/" was interpreted
as cell #3 at the most recently-specified order).  This typically
resulted in spurious tiles near the north pole.
These trailing depth specifiers are optional, but common at least
for MOC 2.0 (where they are required to indicate a resolution
better than the highest explicitly specified order), so this
was probably quite common.  Fix it.

Thanks to Markus Demleitner for spotting this bug.