ttools: fix MOC missing vertex plot bug
There was an error in counting the vertices of a MocVertexData object,
which meant that later vertices were ignored, at a rate of one missed
vertex per cell.  This had the effect of causing MOC cells later in
the list to be missed out or misshapen (if only some of their vertices
were included) during plotting.  The effect is more pronounced for
smaller cells (about the last 20% of vertices would be missed),
since larger (i.e. lower order) ones have their sides interpolated
so they have more than 4 points.

Fixed.