topcat: match case-insensitively for same-name subsets
TopcatModel.addSubset() can either add a new entry to the subset list or
replace an existing one, according to whether the supplied subset has
the "same" name as one already in the list.  At this commit, sameness is
assessed case-insensitively rather than, as previously, case-sensitively.

This is to address an anomaly pointed out by Sergei Klioner, since row subset
referencing in the expression language is done case-insensitively.