ttools: report vounit guesses using later Unity
Upgrade Unity to some 1.1 version and use the new facilities to
report guesswork.  A new UnitStatus.GUESSED_UNIT is introduced.

The Unity version here was downloaded from
https://code.nxg.name/nxg/downloads/unity-1.1-SNAPSHOT-9a4739757561.jar
A 1.1b version is reported to be available soon, and should be a
suitable substitute for this version, but I wanted to get this
functionality in without waiting longer for it.