ttools: upgrade test-time jython version to 2.7.2
The standalone version of JyStilts will in future be distributed with
jython 2.7.2, since the previous version (2.5.3) failed with some
security-related issues when run under Java 17 (though Java 8 and 11
were OK).  So the tests ought to be done using 2.7.2 as well.

The jython package is not used at all for the normal build, but
it is used for testing.  The previous version (jython 2.5.4-rc1)
should be replaced with a version from 2.7.2.

The relevant file is available from the maven central repository:
   https://repo1.maven.org/maven2/org/python/jython-standalone/2.7.2/jython-standalone-2.7.2.jar
It is very large (40Mb), so I don't want to include it in the git
repository if it can be avoided.  Therefore download it at test time
as required.  This approach (downloading dependencies) has not been
used anywhere in starjava up till now as far as I know, all the
dependency jars are committed into the repository.  But in view of
the size of the file, and the stability of the maven central repository,
I think it's reasonable here.  Also this is only required at test time,
not for the build or at runtime.

The file is downloaded into the new directory src/downloads, and only
if the required jar file is not already present.  An 'ant clean' does
not remove it, since it is notionally a dependency rather than a built
file.  It is marked for git-ignoring, so git status won't notice if
it's present or not.  However, we calculate the MD5 checksum when
checking for the file's presence, and the MD5 checksum of the original
download is commited into the repository, so if the downloaded file
becomes corrupted a git status will reveal that.