task: new class MultiChoiceParameter
New concrete subclass of AbstractChoiceParameter which implements
MultiParameter and can supply arrays of the typed option values.

This was introduced for a change that I may end up backing out of,
so it may not be used for now, but it could be useful at some point
in the future, so leave it in place.