ttools: MapEnvironment copes with scalars for arrays
If the value for a parameter with an array type presented to a
MapEnvironment is an instance of the array component type,
the MapEnvironment now turns it into a 1-element array of the
required type, rather than rejecting it as an error.
This is a useful convenience, and aids backward compatibility
if a scalar-typed parameter is changed to an array-typed one.