ttools: upgrade JEL to v2.1.3-pre1
This fixes a long-standing JEL evaluation bug (since JEL 0.9.8)
that provoked a NullPointerException when comparing a string against
a null value; so an expression like "txt==null" would throw a
NullPointerException rather than returning true (or false).
It applied to both literal and evaluated nulls.
That had mostly gone unnoticed, since the NPE generally resulted
in a false value of the whole expression, but (a) that might be
wrong and (b) the exception generation is expensive, so it could
slow things down.

This commit includes a testcase that checks the fix is in place.

I compiled the jel_g.jar from source as usual (ant dist) to get a
version with a suitable classfile version.

jel_g.jar is also copied into the jsky JEL version.
This shouldn't really be necessary, and may not be desirable
(though as far as I know JEL backward compatibility is OK),
but if I don't do it then the topcat unit tests may pick up the
wrong JEL version.  That ought to be fixable instead by adjusting
the test classpath in the topcat build.xml.  The correct solution
would be to put JEL in its own third-party package.

The JEL version included here appears to correspond to JEL savannah
revision 186:

   http://svn.savannah.gnu.org/viewvc/jel?view=revision&revision=186

It was badged as "2.1.3-pre1", but does not have a formal version
number.  Since I believe that Konstantin Metlov, the JEL author,
is based in Ukraine, I don't expect official JEL updates in the near
future.